Initial code for ExoFOP class, new test file to test the class independently
Beginning to think through the ExoFOP class by making an initial call. Using JSON rather than the piped format first to try that out. Created a test file which can be independently run to view all of the JSON data and then below it, what we're picking up in a formatted sense.