Implement orbital period uncertainty in the ExoFOP class
I tried to follow the approach in the NASAExoplanetArchive class here, however, we don't have a positive and negative number in the response from ExoFOP - it's just one float from the field per_e (of which there could be many). So, I copied that result to two variables, then set one of the variables to negative, then left the absolute and square root numpy functions in there. See the comment for what I'll probably replace it with after Kyle reviews this code.